refactor: restructure calendar header into dedicated widget
Separated goal title and month navigation into GoalCalendarHeader
Renamed file and class for clarity
Removed outdated CalendarHeaderSection widget
